requirements.txt
argcomplete==1.10.3
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
Cython==0.29.22
decorator==4.4.2
defusedxml==0.7.1
docutils==0.16
Flask==1.1.2
future==0.18.2
hlm==1.1.9
hst==1.0.0
htf>=2.0
idna==2.10
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jira==2.0.0
junitparser==1.4.1
MarkupSafe==1.1.1
oauthlib==3.1.0
oser==1.3.5
parameterized==0.8.1
paramiko==2.7.2
parso==0.8.1
pbr==5.5.1
pickleshare==0.7.5
ply==3.11
prompt-toolkit==3.0.17
protobuf==3.11.3
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
PyNaCl==1.4.0
PyQRCode==1.2.1
pyserial==3.5
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
termcolor==1.1.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
